# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  products.Product:
    properties:
      brand:
        type: string
      buyNumber:
        type: integer
      categoryKey:
        type: string
      categoryName:
        type: string
      checkCommissionPercent:
        type: number
      commissionPercent:
        type: number
      completeSpec:
        items:
          $ref: '#/definitions/products.csType'
        type: array
      createdAt:
        type: integer
      description:
        type: string
      filterArr:
        items:
          type: string
        type: array
      imageArr:
        items:
          type: string
        type: array
      lowestPrice:
        type: integer
      mainSpecs:
        items:
          $ref: '#/definitions/products.f'
        type: array
      seenNumber:
        type: integer
      spId:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      variationsObj:
        $ref: '#/definitions/products.v'
    type: object
...
  products.csType:
    properties:
      items:
        items:
          $ref: '#/definitions/products.f'
        type: array
      name:
        type: string
    type: object
  products.f:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  products.v:
    properties:
      title:
        type: string
      variations:
        items:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
        type: array
    type: object
...
  similarityGraph.SimilarityNodeRequest:
    properties:
      color:
        type: string
      description:
        type: string
      isCollection:
        type: boolean
      pattern:
        type: string
      productsKeyArray:
        items:
          type: string
        type: array
      public:
        type: boolean
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      userMade:
        type: boolean
    type: object
  similarityGraph.similarityEdge:
    properties:
      _from:
        type: string
      _key:
        type: string
      _to:
        type: string
      coreEdge:
        type: boolean
      type:
        type: string
    required:
    - _from
    - _to
    - type
    type: object
  similarityGraph.similarityNode:
    properties:
      color:
        type: string
      createdAt:
        type: integer
      createdBy:
        type: string
      description:
        type: string
      imageUrl:
        items:
          type: string
        type: array
      isCollection:
        type: boolean
      pattern:
        type: string
      productsKeyArray:
        items:
          type: string
        type: array
      public:
        type: boolean
      seen:
        type: integer
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: integer
      updatedBy:
        type: string
      userKey:
        type: string
      userMade:
        type: boolean
    required:
    - title
    type: object
  similarityGraph.similarityNodeOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      color:
        type: string
      createdAt:
        type: integer
      createdBy:
        type: string
      description:
        type: string
      imageUrl:
        items:
          type: string
        type: array
      isCollection:
        type: boolean
      pattern:
        type: string
      productsKeyArray:
        items:
          type: string
        type: array
      public:
        type: boolean
      seen:
        type: integer
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: integer
      updatedBy:
        type: string
      userKey:
        type: string
      userMade:
        type: boolean
    required:
    - title
    type: object
...
info:
  contact:
    email: armansal1519@gmail.com
    name: Arman Salehi
  description: This is an API for Bamachoub Application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Bamachoub Application
  version: "2.0"
paths:
...
  /p-similarity:
    get:
      consumes:
      - application/json
      description: return all similarity nodes
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/similarityGraph.similarityNode'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: return all similarity nodes
      tags:
      - product similarity
  /p-similarity/{key}:
    delete:
      consumes:
      - application/json
      description: delete similarity node
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: delete similarity node
      tags:
      - product similarity
    get:
      consumes:
      - application/json
      description: return all similar products
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.Product'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: return all similar products
      tags:
      - product similarity
    put:
      consumes:
      - application/json
      description: update similarity node
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/similarityGraph.similarityNode'
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: update similarity node
      tags:
      - product similarity
  /p-similarity/{op}/{productKey}/{nodeKey}:
    post:
      consumes:
      - application/json
      description: add or remove product to a node
      parameters:
      - description: op
        in: path
        name: op
        required: true
        type: string
      - description: productKey
        in: path
        name: productKey
        required: true
        type: string
      - description: nodeKey
        in: path
        name: nodeKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/similarityGraph.similarityNode'
        "404":
          description: Not Found
          schema:
            type: string
      summary: add or remove product to a node
      tags:
      - product similarity
  /p-similarity/edge:
    post:
      consumes:
      - application/json
      description: create similarity edge
      parameters:
      - description: data
        in: body
        name: similarityEdge
        required: true
        schema:
          $ref: '#/definitions/similarityGraph.similarityEdge'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: create similarity edge
      tags:
      - product similarity
  /p-similarity/graph-sim/{key}:
    get:
      consumes:
      - application/json
      description: return all similar nodes
      parameters:
      - description: similarity type
        in: query
        name: type
        type: string
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/similarityGraph.similarityNodeOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: return all similar nodes
      tags:
      - product similarity
  /p-similarity/node:
    post:
      consumes:
      - application/json
      description: create similarity node
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/similarityGraph.SimilarityNodeRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create similarity node
      tags:
      - product similarity
  /p-similarity/node/{key}:
    get:
      consumes:
      - application/json
      description: return one node with products
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.Product'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: return one node with products
      tags:
      - product similarity
  /p-similarity/node/admin:
    post:
      consumes:
      - application/json
      description: create similarity node by admin
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/similarityGraph.SimilarityNodeRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: create similarity node by admin
      tags:
      - product similarity
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"